Most Recent Responsive Additions
I have been fiddling with several aspects of CSS properties. I had some trouble with the footer element. There was a small block of uncovered space at the bottom of the page. Once I found the CSS left property, I was able to get that block to go away. Instead of using CSS on a <footer>, I used it on a <div> and gave it a "left: 0".