Finally Pushing With A Title From The CLI
Like the title says, I have finally gotten some understanding about this git thing, and I'm pushing from the CLI. Of course, that's not a big deal to anyone with a lot of experience, a few simple commands, but to me (someone with very little coding experience) it gives me a sense of accomplishment. Yay!