Organizational Prep
I dug in to organizing the HTML comments for this document, and polished the media queries for 3 screen widths. I added some accesibility considerations. I adjusted some things for the sake of aesthetics. Next I want to optimize the aesthetics to be a little less generic, and a little bit more hip. I'll need to find a font package and a color scheme. I'll use Adobe products to find a color scheme. Perhaps I'll include some SVGs or CSS graphics.