Github Copilot and Chat GPT join me
So I took the plunge yesterday, and I'm totally blown away by the power of these artificial intelligence assistants. There is no going back. I will be continuing to learn about this stuff--but AI is absolutely on board. I made some tweaks to the structure and CSS of the blog page for the  Responsive Newsletter Project.